03-03-23

+++ libft/ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: atardif <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/07 14:49:30 by atardif           #+#    #+#             */
-/*   Updated: 2023/02/23 11:45:15 by atardif          ###   ########.fr       */
+/*   Updated: 2023/03/03 17:29:24 by atardif          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ long	ft_atoi(const char *nptr)
 {
 	int	i;
 	int	sign;
-	int	result;
+	long	result;
 
 	i = 0;
 	sign = 1;
